removes mnemonic pop-up in watcher table
