--- src/Command/Components/Watcher/MnemonicPopUp.css
@@ -1,18 +1,0 @@
-.mnemonic-pop-up rux-card span {
-  margin-inline: var(--spacing-1);
-}
-
-.mnemonic-pop-up rux-icon {
-  margin-top: var(--spacing-1);
-  margin-left: var(--spacing-1);
-  padding: 0;
-  margin-right: 0;
-}
-
-.mnemonic-pop-up rux-button {
-  margin-left: var(--spacing-4);
-}
-
-.mnemonic-pop-up rux-button::part(container) {
-  padding: 0;
-}
