+++ src/Command/Components/Watcher/Watcher.css
@@ -49,33 +49,12 @@
   transform: rotate(-45deg);
 }
 
-/* MNEMONIC POP-UP STYLES */
-
-rux-pop-up.mnemonic-pop-up span[slot="trigger"] {
+rux-tooltip {
   color: var(--color-text-interactive-default);
   display: inline-block;
+  border-bottom: 1px solid transparent;
 }
 
-rux-pop-up.mnemonic-pop-up span[slot="trigger"]:hover {
-  box-shadow: inset 0 -1px 0 0 var(--color-text-interactive-default);
-}
-
-rux-pop-up.mnemonic-pop-up rux-card::part(container) {
-  border-width: 0px;
-  box-shadow: none;
-}
-
-rux-pop-up.mnemonic-pop-up rux-card::part(body) {
-  display: flex;
-  flex-direction: column;
-}
-
-rux-pop-up.mnemonic-pop-up rux-card div {
-  display: flex;
-  justify-content: space-between;
-  gap: var(--spacing-8);
-}
-
-rux-pop-up.mnemonic-pop-up rux-checkbox::part(label) {
-  font-weight: var(--font-weights-bold);
+rux-tooltip:hover {
+  border-bottom: 1px solid var(--color-text-interactive-default);
 }
