+++ src/Command/Components/Watcher/Watcher.tsx
@@ -76,8 +76,12 @@ const Watcher = () => {
               </RuxTableHeaderCell>
               <RuxTableHeaderCell>Mnemonic</RuxTableHeaderCell>
               <RuxTableHeaderCell>Unit</RuxTableHeaderCell>
-              <RuxTableHeaderCell>Threshold</RuxTableHeaderCell>
-              <RuxTableHeaderCell>Actual</RuxTableHeaderCell>
+              <RuxTableHeaderCell className="text-align-right">
+                Threshold
+              </RuxTableHeaderCell>
+              <RuxTableHeaderCell className="text-align-right">
+                Actual
+              </RuxTableHeaderCell>
               <RuxTableHeaderCell>
                 {/* placeholder for actions menu column */}
               </RuxTableHeaderCell>
